requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
blinker==1.7.0
Bottleneck==1.3.8
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.96.1
filelock==3.13.3
fonttools==4.25.0
frozenlist==1.4.1
fsspec==2024.2.0
gmpy2==2.1.5
google-auth==2.29.0
google-auth-oauthlib==1.2.0
grpcio==1.62.1
h11==0.14.0
huggingface-hub==0.21.4
idna==3.6
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munkres==1.1.4
nest-asyncio==1.6.0
networkx==3.3
numexpr==2.10.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.0
ply==3.11
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==1.10.14
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-crf==0.7.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.11.4
seaborn==0.13.2
sip==6.8.3
six==1.16.0
sniffio==1.3.1
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.4.0
tokenizers==0.15.2
tomli==2.0.1
torch==2.1.2
torchaudio==2.1.2
torchmetrics==1.3.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.22.0
wcwidth==0.2.13
Werkzeug==3.0.2
win-inet-pton==1.1.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
streamlit